[FEATURE IMPL] 3 of 4 functions for Hands-on (Persistent storage is missing)
New Feature:
This new feature set consists of 3 main function.
The first commit realize the ZC (Light) and ZR (Switch) basic functions as forming and joining to a network.
The second commit includes the implementation of ON/OFF command by pressing the Button1 of the router. The appropriate callback is enabled on the Coordinator side, and the LED setting happens there.
The third feature is a 1 second period LED blinking event on the router. This commit affects only the Switch app.