[DOC IMP] Update the contributing.md and pull request template
1.
Use dash line '-' if multiple words present in the folder or documentation name
2.
Update the documentation rule to emphasize that a files folder is needed to organize all of the attachments except markdown.
3.
Mention that there is a dedicated wiki repo for documentation maintenance, and how to contribute documentations.
4.
Add screenshot to illustrate the required directory structure
5.
Update the history information
6.
Update the pull request template